fixed New Project Dialog
